backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.25.2
python-socketio==5.10.0
pydantic==2.12.5
pydantic-settings==2.4.0
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
apscheduler==3.10.4
python-multipart==0.0.6
xgboost==2.0.3
scikit-learn==1.3.2
lightgbm==4.1.0
joblib==1.3.2

# Protobuf support for Upstox V3
protobuf==5.28.0
grpcio==1.66.0
grpcio-tools==1.66.0
